@@ -4,21 +4,28 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
 {
     private readonly DownloadConfiguration _config = new() {
         Timeout = 100,
-        MaxTryAgainOnFailover = 100,
+        MaxTryAgainOnFailure = 100,
         BufferBlockSize = 1024
     };
 
     [Theory]
-    [InlineData(-1, 1024)] // Chunk File By Negative Parts Test
-    [InlineData(0, 1024)] // Chunk File By Zero Parts Test
     [InlineData(1, 1024)] // Chunk File Positive 1 Parts Test
     public void SingleChunkFileTest(int chunkCount, long fileSize)
     {
         // act 
-        using var package = ChunkFileTest(chunkCount, fileSize);
+        using DownloadPackage package = ChunkFileTest(chunkCount, fileSize);
 
         // assert
         Assert.Single(package.Chunks);
+    }
+    
+    [Theory]
+    [InlineData(-10, 1024)] // Chunk File By Negative Parts Test
+    [InlineData(-1, 1024)] // Chunk File By Negative Parts Test
+    [InlineData(0, 1024)] // Chunk File By Zero Parts Test
+    public void ChunkFileWithErrorTest(int chunkCount, long fileSize)
+    {
+        Assert.Throws<ArgumentOutOfRangeException>(() => ChunkFileTest(chunkCount, fileSize));
     }
 
     [Theory]
@@ -29,7 +36,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
     public void PositiveChunkFileTest(int chunkCount, long fileSize)
     {
         // act 
-        using var package = ChunkFileTest(chunkCount, fileSize);
+        using DownloadPackage package = ChunkFileTest(chunkCount, fileSize);
 
         // assert
         Assert.Equal(chunkCount, package.Chunks.Length);
@@ -41,7 +48,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
     public void ChunkFileMoreThanSizeTest(int chunkCount, long fileSize)
     {
         // act 
-        using var package = ChunkFileTest(chunkCount, fileSize);
+        using DownloadPackage package = ChunkFileTest(chunkCount, fileSize);
 
         // assert
         Assert.Equal(fileSize, package.Chunks.Length);
@@ -57,7 +64,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
         long totalBytes = _config.RangeHigh - _config.RangeLow + 1;
 
         // act
-        using var package = ChunkFileTest(64, totalBytes);
+        using DownloadPackage package = ChunkFileTest(64, totalBytes);
 
         // assert
         Assert.Equal(totalBytes, package.Chunks.Sum(chunk => chunk.Length));
@@ -74,7 +81,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
         long actualTotalSize = _config.RangeHigh + 1;
 
         // act
-        using var package = ChunkFileTest(64, actualTotalSize);
+        using DownloadPackage package = ChunkFileTest(64, actualTotalSize);
 
         // assert
         Assert.Equal(actualTotalSize, package.Chunks.Sum(chunk => chunk.Length));
@@ -86,7 +93,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
     public void ChunkFileZeroSizeTest()
     {
         // act
-        using var package = ChunkFileTest(64, 0);
+        using DownloadPackage package = ChunkFileTest(64, 0);
 
         // assert
         Assert.Single(package.Chunks);
@@ -103,7 +110,7 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
     public void PositiveChunkFileRangeTest(int chunkCount, long fileSize)
     {
         // act 
-        using var package = ChunkFileTest(chunkCount, fileSize);
+        using DownloadPackage package = ChunkFileTest(chunkCount, fileSize);
 
         // assert
         Assert.Equal(0, package.Chunks[0].Start);
@@ -118,10 +125,10 @@ public class ChunkHubTest(ITestOutputHelper output) : BaseTestClass(output)
     private DownloadPackage ChunkFileTest(int chunkCount, long fileSize = 1024)
     {
         // arrange
-        var package = new DownloadPackage {
+        DownloadPackage package = new() {
             TotalFileSize = fileSize
         };
-        var chunkHub = new ChunkHub(_config);
+        ChunkHub chunkHub = new(_config);
 
         // act
         _config.ChunkCount = chunkCount;
