@@ -7,17 +7,17 @@ public partial class Program
 {
     private static DownloadConfiguration GetDownloadConfiguration()
     {
-        var cookies = new CookieContainer();
+        CookieContainer cookies = new();
         cookies.Add(new Cookie("download-type", "test") { Domain = "domain.com" });
 
         return new DownloadConfiguration {
             BufferBlockSize = 10240,    // usually, hosts support max to 8000 bytes, default values is 8000
             ChunkCount = 8,             // file parts to download, default value is 1
+            ParallelCount = 4,          // number of parallel downloads. The default value is the same as the chunk count
             MaximumBytesPerSecond = 1024 * 1024 * 20,  // download speed limited to 20MB/s, default values is zero or unlimited
-            MaxTryAgainOnFailover = 50_000,  // the maximum number of times to fail
+            MaxTryAgainOnFailure = 50_000,  // the maximum number of times to fail
             MaximumMemoryBufferBytes = 1024 * 1024 * 500, // release memory buffer after each 500MB
             ParallelDownload = true,    // download parts of file as parallel or not. Default value is false
-            ParallelCount = 8,          // number of parallel downloads. The default value is the same as the chunk count
             Timeout = 3000,             // timeout (millisecond) per stream block reader, default value is 1000
             RangeDownload = false,      // set true if you want to download just a specific range of bytes of a large file
             RangeLow = 0,               // floor offset of download range of a large file
