@@ -7,14 +7,13 @@ global using System.Collections.Generic;
 global using System.IO;
 global using System.Threading;
 global using System.Threading.Tasks;
-global using Downloader.Test.Helper;
-global using Microsoft.Extensions.Logging;
 global using System.ComponentModel;
 global using System.Diagnostics;
 global using System.Net;
 global using System.Text;
 global using System.Net.Http;
+global using System.Reflection;
+global using Microsoft.Extensions.Logging;
+global using Newtonsoft.Json;
 global using Downloader.Test.Helper;
 global using Downloader.Extensions.Helpers;
-global using Newtonsoft.Json;
-global using System.Reflection;
