@@ -4,13 +4,13 @@ public class ChunkDownloaderOnFileTest : ChunkDownloaderTest
 {
     public ChunkDownloaderOnFileTest(ITestOutputHelper output) : base(output)
     {
-        var path = Path.GetTempFileName();
+        string path = Path.GetTempFileName();
         Configuration = new DownloadConfiguration {
             BufferBlockSize = 1024,
             ChunkCount = 16,
             ParallelDownload = true,
             ParallelCount = 8,
-            MaxTryAgainOnFailover = 100,
+            MaxTryAgainOnFailure = 100,
             MinimumSizeOfChunking = 16,
             Timeout = 100,
         };
