@@ -22,7 +22,7 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
         await Storage.FlushAsync();
 
         // act
-        var reader = Storage.OpenRead();
+        Stream reader = Storage.OpenRead();
 
         // assert
         Assert.Equal(DataLength, reader.Length);
@@ -37,7 +37,7 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
         await Storage.FlushAsync();
 
         // act
-        var reader = Storage.OpenRead();
+        Stream reader = Storage.OpenRead();
 
         // assert
         for (int i = 0; i < DataLength; i++)
@@ -50,8 +50,8 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     public async Task SlowWriteTest()
     {
         // arrange
-        var data = new byte[] { 1 };
-        var size = 1024;
+        byte[] data = new byte[] { 1 };
+        int size = 1024;
         CreateStorage(size);
 
         // act
@@ -59,7 +59,7 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
             await Storage.WriteAsync(i, data, 1);
 
         await Storage.FlushAsync();
-        var readerStream = Storage.OpenRead();
+        Stream readerStream = Storage.OpenRead();
 
         // assert
         Assert.Equal(size, Storage.Length);
@@ -72,7 +72,7 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(0);
-        var length = DataLength / 2;
+        int length = DataLength / 2;
 
         // act
         await Storage.WriteAsync(0, Data, length);
@@ -87,12 +87,12 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(DataLength);
-        var length = DataLength / 2;
+        int length = DataLength / 2;
 
         // act
         await Storage.WriteAsync(0, Data, length);
         await Storage.FlushAsync();
-        var reader = Storage.OpenRead();
+        Stream reader = Storage.OpenRead();
 
         // assert
         for (int i = 0; i < length; i++)
@@ -106,19 +106,19 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(DataLength);
-        var count = 128;
-        var size = DataLength / count;
+        int count = 128;
+        int size = DataLength / count;
 
         // act
         for (int i = 0; i < count; i++)
         {
-            var startOffset = i * size;
+            int startOffset = i * size;
             await Storage.WriteAsync(startOffset, Data.Skip(startOffset).Take(size).ToArray(), size);
         }
         await Storage.FlushAsync();
 
         // assert
-        var reader = Storage.OpenRead();
+        Stream reader = Storage.OpenRead();
         for (int i = 0; i < DataLength; i++)
         {
             Assert.Equal(Data[i], reader.ReadByte());
@@ -130,10 +130,10 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(0);
-        var length = DataLength + 1;
+        int length = DataLength + 1;
 
         // act
-        var writeMethod = async () => await Storage.WriteAsync(0, Data, length);
+        Func<Task> writeMethod = async () => await Storage.WriteAsync(0, Data, length);
 
         // assert
         await Assert.ThrowsAnyAsync<ArgumentOutOfRangeException>(writeMethod);
@@ -172,12 +172,12 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(0);
-        var data = new byte[] { 0x0, 0x1, 0x2, 0x3, 0x4 };
+        byte[] data = new byte[] { 0x0, 0x1, 0x2, 0x3, 0x4 };
         await Storage.WriteAsync(0, data, 1);
         await Storage.FlushAsync();
 
         // act
-        var actualLength = Storage.Length;
+        long actualLength = Storage.Length;
 
         // assert
         Assert.Equal(1, actualLength);
@@ -188,14 +188,14 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     {
         // arrange
         CreateStorage(0);
-        var data = new byte[] { 0x0, 0x1, 0x2, 0x3, 0x4 };
+        byte[] data = new byte[] { 0x0, 0x1, 0x2, 0x3, 0x4 };
         await Storage.WriteAsync(0, data, data.Length);
         await Storage.FlushAsync();
 
         // act
-        var serializedStream = JsonConvert.SerializeObject(Storage);
+        string serializedStream = JsonConvert.SerializeObject(Storage);
         Storage.Dispose();
-        using var mutableStream = JsonConvert.DeserializeObject<ConcurrentStream>(serializedStream);
+        using ConcurrentStream mutableStream = JsonConvert.DeserializeObject<ConcurrentStream>(serializedStream);
         await mutableStream.WriteAsync(mutableStream.Position, data, data.Length);
         await mutableStream.FlushAsync();
 
@@ -207,26 +207,26 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     public async Task TestDynamicBufferData()
     {
         // arrange
-        var size = 1024; // 1KB
+        int size = 1024; // 1KB
         CreateStorage(size);
 
         // act
         for (int i = 0; i < size / 8; i++)
         {
-            var data = new byte[10]; // zero bytes
+            byte[] data = new byte[10]; // zero bytes
             Array.Fill(data, (byte)i);
             await Storage.WriteAsync(i * 8, data, 8);
         }
         await Storage.FlushAsync();
-        var readerStream = Storage.OpenRead();
+        Stream readerStream = Storage.OpenRead();
 
         // assert
         Assert.Equal(size, Storage.Length);
         for (int i = 0; i < size / 8; i++)
         {
-            var data = new byte[8]; // zero bytes
+            byte[] data = new byte[8]; // zero bytes
             Array.Fill(data, (byte)i);
-            var buffer = new byte[8];
+            byte[] buffer = new byte[8];
             Assert.Equal(8, readerStream.Read(buffer, 0, 8));
             Assert.True(buffer.SequenceEqual(data));
         }
@@ -238,17 +238,17 @@ public abstract class StorageTest(ITestOutputHelper output) : BaseTestClass(outp
     public async Task TestSerialization()
     {
         // arrange
-        var size = 256;
-        var data = DummyData.GenerateOrderedBytes(size);
+        int size = 256;
+        byte[] data = DummyData.GenerateOrderedBytes(size);
         CreateStorage(size);
         await Storage.WriteAsync(0, data, size);
         await Storage.FlushAsync();
 
         // act
-        var serializedStream = JsonConvert.SerializeObject(Storage);
+        string serializedStream = JsonConvert.SerializeObject(Storage);
         Storage.Dispose();
-        using var newStream = JsonConvert.DeserializeObject<ConcurrentStream>(serializedStream);
-        var readerStream = newStream.OpenRead();
+        using ConcurrentStream newStream = JsonConvert.DeserializeObject<ConcurrentStream>(serializedStream);
+        Stream readerStream = newStream.OpenRead();
 
         // assert
         Assert.Equal(size, readerStream.Length);
