@@ -13,7 +13,7 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     private int _bufferBlockSize = 1024; // usually, hosts support max to 8000 bytes
     private int _chunkCount = 1; // file parts to download
     private long _maximumBytesPerSecond = ThrottledStream.Infinite; // No-limitation in download speed
-    private int _maximumTryAgainOnFailover = int.MaxValue; // the maximum number of times to fail.
+    private int _maximumTryAgainOnFailure = int.MaxValue; // the maximum number of times to fail.
     private long _maximumMemoryBufferBytes;
     private bool _checkDiskSizeBeforeDownload = true; // check disk size for temp and file path
     private bool _parallelDownload; // download parts of file as parallel or not
@@ -23,29 +23,33 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     private long _rangeLow; // starting byte offset
     private long _rangeHigh; // ending byte offset
 
-    private bool
-        _clearPackageOnCompletionWithFailure; // Clear package and downloaded data when download completed with failure
+    // Clear package and downloaded data when download completed with failure
+    private bool _clearPackageOnCompletionWithFailure;
 
-    private long _minimumSizeOfChunking = 512; // minimum size of chunking to download a file in multiple parts
+    // minimum size of chunking to download a file in multiple parts
+    private long _minimumSizeOfChunking = 512;
 
-    private bool
-        _reserveStorageSpaceBeforeStartingDownload; // Before starting the download, reserve the storage space of the file as file size.
+    // Before starting the download, reserve the storage space of the file as file size.
+    private bool _reserveStorageSpaceBeforeStartingDownload;
 
-    private bool
-        _enableLiveStreaming; // Get on demand downloaded data with ReceivedBytes on downloadProgressChanged event 
+    // Get on demand downloaded data with ReceivedBytes on downloadProgressChanged event
+    private bool _enableLiveStreaming;
 
     /// <summary>
     /// To bind view models to fire changes in MVVM pattern
     /// </summary>
     public event PropertyChangedEventHandler PropertyChanged = delegate { };
-    
-    private void OnPropertyChanged<T>(ref T field, T newValue, [CallerMemberName] string name = null)
+
+    /// <summary>
+    /// Notify every change of configuration properties
+    /// </summary>
+    protected virtual void OnPropertyChanged<T>(ref T field, T newValue, [CallerMemberName] string name = null)
     {
-        if (!field.Equals(newValue))
-        {
-            field = newValue;
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-        }
+        if (field.Equals(newValue))
+            return;
+
+        field = newValue;
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
     }
 
     /// <summary>
@@ -63,7 +67,13 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public int BufferBlockSize
     {
         get => (int)Math.Min(MaximumSpeedPerChunk, _bufferBlockSize);
-        set => OnPropertyChanged(ref _bufferBlockSize, value);
+        set
+        {
+            if (value is < 1 or > 1048576) // 1MB = 1024 * 1024 bytes
+                throw new ArgumentOutOfRangeException(nameof(BufferBlockSize),
+                    "Buffer block size must be between 1 byte and 1024KB");
+            OnPropertyChanged(ref _bufferBlockSize, value);
+        }
     }
 
     /// <summary>
@@ -81,7 +91,12 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public int ChunkCount
     {
         get => _chunkCount;
-        set => OnPropertyChanged(ref _chunkCount, Math.Max(1, value));
+        set
+        {
+            if (value < 1)
+                throw new ArgumentOutOfRangeException(nameof(ChunkCount), "Chunk count must be greater than 0");
+            OnPropertyChanged(ref _chunkCount, Math.Max(1, value));
+        }
     }
 
     /// <summary>
@@ -90,24 +105,29 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public long MaximumBytesPerSecond
     {
         get => _maximumBytesPerSecond;
-        set => OnPropertyChanged(ref _maximumBytesPerSecond, value <= 0 ? long.MaxValue : value);
+        set
+        {
+            if (value < 0)
+                throw new ArgumentOutOfRangeException(nameof(MaximumBytesPerSecond),
+                    "Maximum bytes per second cannot be negative");
+            OnPropertyChanged(ref _maximumBytesPerSecond, value <= 0 ? long.MaxValue : value);
+        }
     }
 
     /// <summary>
     /// Gets the maximum bytes per second that can be transferred through the base stream at each chunk downloader.
     /// This property is read-only.
     /// </summary>
-    public long MaximumSpeedPerChunk => ParallelDownload
-        ? MaximumBytesPerSecond / Math.Max(Math.Min(Math.Min(ChunkCount, ParallelCount), ActiveChunks), 1)
-        : MaximumBytesPerSecond;
+    public long MaximumSpeedPerChunk => MaximumBytesPerSecond /
+                                        Math.Max(Math.Min(Math.Min(ChunkCount, ParallelCount), ActiveChunks), 1);
 
     /// <summary>
     /// Gets or sets the maximum number of times to try again to download on failure.
     /// </summary>
-    public int MaxTryAgainOnFailover
+    public int MaxTryAgainOnFailure
     {
-        get => _maximumTryAgainOnFailover;
-        set => OnPropertyChanged(ref _maximumTryAgainOnFailover, value);
+        get => _maximumTryAgainOnFailure;
+        set => OnPropertyChanged(ref _maximumTryAgainOnFailure, value);
     }
 
     /// <summary>
@@ -153,7 +173,13 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public long RangeHigh
     {
         get => _rangeHigh;
-        set => OnPropertyChanged(ref _rangeHigh, value);
+        set
+        {
+            if (value < 0)
+                throw new ArgumentOutOfRangeException(nameof(RangeHigh),
+                    "Range high cannot be negative");
+            OnPropertyChanged(ref _rangeHigh, value);
+        }
     }
 
     /// <summary>
@@ -167,7 +193,13 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public int Timeout
     {
         get => _timeout;
-        set => OnPropertyChanged(ref _timeout, value);
+        set
+        {
+            if (value < 100)
+                throw new ArgumentOutOfRangeException(nameof(Timeout),
+                    "Timeout must be at least 100 milliseconds");
+            OnPropertyChanged(ref _timeout, value);
+        }
     }
 
     /// <summary>
@@ -202,11 +234,22 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     /// Gets or sets the maximum amount of memory, in bytes, that the Downloader library is allowed
     /// to allocate for buffering downloaded content. Once this limit is reached, the library will
     /// stop downloading and start writing the buffered data to a file stream before continuing.
-    /// The default value for is 0, which indicates unlimited buffering.
+    /// The default value is 0, which indicates unlimited buffering.
+    /// 
+    /// This setting is particularly useful when:
+    /// 1. Downloading large files on systems with limited memory
+    /// 2. Preventing out-of-memory exceptions during parallel downloads
+    /// 3. Optimizing memory usage for long-running download operations
+    /// 
+    /// Recommended values:
+    /// - For systems with 4GB RAM: 256MB (268435456 bytes)
+    /// - For systems with 8GB RAM: 512MB (536870912 bytes)
+    /// - For systems with 16GB RAM: 1GB (1073741824 bytes)
+    /// 
+    /// Note: Setting this value too low may impact download performance due to frequent disk I/O operations.
     /// </summary>
     /// <example>
-    /// The following example sets the maximum memory buffer to 50 MB, causing the library to release
-    /// the memory buffer after each 50 MB of downloaded content:
+    /// The following example sets the maximum memory buffer to 50 MB:
     /// <code>
     /// MaximumMemoryBufferBytes = 1024 * 1024 * 50
     /// </code>
@@ -214,13 +257,26 @@ public class DownloadConfiguration : ICloneable, INotifyPropertyChanged
     public long MaximumMemoryBufferBytes
     {
         get => _maximumMemoryBufferBytes;
-        set => OnPropertyChanged(ref _maximumMemoryBufferBytes, value);
+        set
+        {
+            if (value < 0)
+                throw new ArgumentOutOfRangeException(nameof(MaximumMemoryBufferBytes),
+                    "Maximum memory buffer bytes cannot be negative");
+            OnPropertyChanged(ref _maximumMemoryBufferBytes, value);
+        }
     }
 
     /// <summary>
     /// Gets or sets a value indicating whether live-streaming is enabled or not. If it's enabled, get the on-demand downloaded data
     /// with ReceivedBytes on the downloadProgressChanged event.
-    /// Note: This option may consume more memory because it copies each block of downloaded data into ReceivedBytes.
+    /// 
+    /// Important considerations:
+    /// 1. Enabling this option will increase memory usage as each downloaded block is copied to the ReceivedBytes buffer
+    /// 2. The memory impact is proportional to the BufferBlockSize and download speed
+    /// 3. For large files, consider setting MaximumMemoryBufferBytes to limit memory usage
+    /// 4. This feature is best used when you need to process downloaded data in real-time
+    /// 
+    /// Default value is false.
     /// </summary>
     public bool EnableLiveStreaming
     {
