@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net;
+using System.Net.Http;
 using System.Net.Security;
 using System.Net.Sockets;
 using System.Security.Cryptography.X509Certificates;
@@ -8,15 +9,50 @@ namespace Downloader.Extensions.Helpers;
 
 internal static class ExceptionHelper
 {
+    internal static bool IsRedirectError(this Exception error)
+    {
+        return error is HttpRequestException { StatusCode: not null } responseException &&
+               IsRedirectStatus(responseException.StatusCode.Value);
+    }
+
+    internal static bool IsRedirectStatus(this HttpStatusCode statusCode)
+    {
+        return statusCode is
+            HttpStatusCode.Moved or
+            HttpStatusCode.Redirect or
+            HttpStatusCode.RedirectMethod or
+            HttpStatusCode.TemporaryRedirect or
+            HttpStatusCode.PermanentRedirect;
+    }
+
+    internal static bool IsRequestedRangeNotSatisfiable(this Exception error)
+    {
+        return error is HttpRequestException { StatusCode: HttpStatusCode.RequestedRangeNotSatisfiable };
+    }
+
     internal static bool IsMomentumError(this Exception error)
     {
-        if (error.HasSource("System.Net.Http", "System.Net.Sockets", "System.Net.Security"))
-            return true;
+        if (error is HttpRequestException {
+                StatusCode: HttpStatusCode.InternalServerError or
+                HttpStatusCode.BadGateway
+            })
+            return false;
 
-        if (error.HasTypeOf(typeof(WebException), typeof(SocketException)))
+        if (error is HttpRequestException {
+                StatusCode: HttpStatusCode.Ambiguous or
+                HttpStatusCode.TooManyRequests or
+                HttpStatusCode.ServiceUnavailable or
+                HttpStatusCode.GatewayTimeout
+            })
+        {
             return true;
+        }
 
-        return false;
+        var result = error.IsRedirectError() ||
+                     error.HasSource("System.Net.Http", "System.Net.Sockets", "System.Net.Security") ||
+                     error.HasTypeOf(typeof(WebException), typeof(SocketException));
+
+        return result;
     }
 
     internal static bool HasTypeOf(this Exception exp, params Type[] types)
@@ -57,13 +93,10 @@ internal static class ExceptionHelper
     /// Sometime a server get certificate validation error
     /// https://stackoverflow.com/questions/777607/the-remote-certificate-is-invalid-according-to-the-validation-procedure-using
     /// </summary>
-    /// <param name="sender"></param>
-    /// <param name="certificate"></param>
-    /// <param name="chain"></param>
-    /// <param name="sslPolicyErrors"></param>
     internal static bool CertificateValidationCallBack(object sender,
         X509Certificate certificate,
-        X509Chain chain, SslPolicyErrors sslPolicyErrors)
+        X509Chain chain,
+        SslPolicyErrors sslPolicyErrors)
     {
         // If the certificate is a valid, signed certificate, return true.
         if (sslPolicyErrors == SslPolicyErrors.None)
@@ -101,10 +134,8 @@ internal static class ExceptionHelper
             // for default Exchange server installations, so return true.
             return true;
         }
-        else
-        {
-            // In all other cases, return false.
-            return false;
-        }
+
+        // In all other cases, return false.
+        return false;
     }
 }
