@@ -13,7 +13,7 @@ public class PauseTokenSource
     /// <summary>
     /// Gets the pause token associated with this source.
     /// </summary>
-    public PauseToken Token => new PauseToken(this);
+    public PauseToken Token => new(this);
 
     /// <summary>
     /// Gets a value indicating whether the operation is paused.
@@ -40,7 +40,7 @@ public class PauseTokenSource
         // and the time we did the compare-exchange, repeat.
         while (true)
         {
-            var tcs = _tcsPaused;
+            TaskCompletionSource<bool> tcs = _tcsPaused;
 
             if (tcs == null)
                 return;
